+++ internal/provider/sql_user_resource_test.go
@@ -70,6 +70,7 @@ func TestIntegrationSqlUserResource(t *testing.T) {
 		Config: client.ClusterConfig{
 			Serverless: &client.ServerlessClusterConfig{
 				RoutingId: "routing-id",
+				UpgradeType: client.UPGRADETYPETYPE_AUTOMATIC,
 			},
 		},
 		State: "CREATED",